# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My OpenApi Documentation",
        "description": "My OpenApi description",
        "contact": {
            "email": "markusmak@gmail.com"
        },
        "license": {
            "name": "Proprietary",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "My API Server"
        }
    ],
    "paths": {
...
        "/api/invoice": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Exibe uma lista dos recursos de faturas.",
                "operationId": "def0c3217a818e2d400c45cece79781c",
                "responses": {
                    "200": {
                        "description": "Retorna a lista de faturas com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se ocorrer um problema durante a busca das faturas."
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Armazena um novo recurso de fatura.",
                "operationId": "5419d250afff57c940145ce442f3cc57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna os dados da nova fatura criada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se ocorrer um problema durante a criação da fatura."
                    }
                }
            }
        },
        "/api/invoice/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Exibe o recurso de fatura especificado.",
                "operationId": "37401f316ec16144ce12c3c441b8083d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fatura a ser recuperada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da fatura encontrada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se a fatura não for encontrada."
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Atualiza o recurso de fatura especificado.",
                "operationId": "9a050ae92eaf55696d2ec4225feed57a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fatura a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da fatura atualizada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se a fatura não for encontrada."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Exclui o recurso de fatura especificado.",
                "operationId": "697dce9813c3870ea6d416d2a82f5290",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fatura a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "force",
                        "in": "query",
                        "description": "Força a deleção",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma mensagem de sucesso após a exclusão."
                    },
                    "400": {
                        "description": "Retorna um erro se a fatura não for encontrada."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoiceRequest": {
                "title": "InvoiceRequest",
                "description": "Request schema for Invoice",
                "required": [
                    "cadastre_id",
                    "signature_id",
                    "describe",
                    "value",
                    "expiration"
                ],
                "properties": {
                    "cadastre_id": {
                        "description": "ID of the related cadastre",
                        "type": "integer",
                        "example": 1
                    },
                    "signature_id": {
                        "description": "ID of the related signature",
                        "type": "integer",
                        "example": 1
                    },
                    "describe": {
                        "description": "Description of the invoice",
                        "type": "string",
                        "example": "Invoice description"
                    },
                    "value": {
                        "description": "Value of the signature",
                        "type": "string",
                        "example": "Signature value"
                    },
                    "expiration": {
                        "description": "Expiration of the signature",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    }
                },
                "type": "object"
            },
...
        }
    }
}